+++ src/Models/Product.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace VaniloCloud\Models;
 
+use VaniloCloud\Enums\ProductState;
+
 class Product
 {
     use HasId;
@@ -17,7 +19,7 @@ class Product
     public readonly ?float $backorder;
     public readonly ?float $price;
     public readonly ?float $original_price;
-    public readonly string $state;
+    public readonly ?ProductState $state;
     public readonly string $tax_category_id;
     public readonly ?string $description;
     public readonly ?string $short_description;